fix: 일반 api 함수 로직 분리 및 명명법 변경
