+++ src/apis/nickNameCheck.ts
@@ -1,13 +1,15 @@
-import client from "@/apis";
+import client from "@/apis/client";
 
 export type NickNameCheckResponse = {
   code: string;
   message: string;
 };
 
-export const nickNameCheck = async (nickName: string) => {
+const nickNameCheck = async (nickName: string) => {
   const response = await client.get<NickNameCheckResponse>(
     `/members/check-nickname?nickname=${nickName}`
   );
   return response.data;
 };
+
+export default nickNameCheck;
