+++ src/libraries/reactQuery/ReactQuerySetting.tsx
@@ -1,13 +1,13 @@
-import { PropsWithChildren } from 'react';
+import { PropsWithChildren } from "react";
 
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 
 const MINUTES = 1000 * 60;
 
 const queryClient = new QueryClient({
   defaultOptions: {
     queries: {
-      retry: 1,
+      retry: false,
       retryDelay: 1000,
       staleTime: 5 * MINUTES,
     },
@@ -15,5 +15,7 @@ const queryClient = new QueryClient({
 });
 
 export default function ReactQuerySetting({ children }: PropsWithChildren) {
-  return <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>;
+  return (
+    <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+  );
 }
